Rename async_data_handler
