+++ game/terrain/__init__.py
@@ -1,5 +1,5 @@
 from .terrain_handler import *
-from .async_data_handler import *
+from .data_handler import *
 from .chunk import *
 from .tile import *
